+++ src/main/java/lagatrix/client/gui/window/MainWindow.java
@@ -8,12 +8,12 @@ import java.awt.Color;
  * @author javierfh03
  * @since 0.1
  */
-public class WindowMain extends javax.swing.JFrame {
+public class MainWindow extends javax.swing.JFrame {
 
     /**
      * Constructor of the class.
      */
-    public WindowMain() {
+    public MainWindow() {
         setUndecorated(true);
         initComponents();
         setBackground(new Color(0.0F, 0.0F, 0.0F, 0.0F));
@@ -33,7 +33,6 @@ public class WindowMain extends javax.swing.JFrame {
         header = new lagatrix.client.gui.components.complex.fragment.Header();
         menu = new lagatrix.client.gui.components.complex.fragment.Menu();
         viewsPanel = new javax.swing.JPanel();
-        monitoringView1 = new lagatrix.client.gui.views.main.MonitoringView();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setTitle("Lagatrix");
@@ -46,7 +45,6 @@ public class WindowMain extends javax.swing.JFrame {
         viewsPanel.setBackground(new java.awt.Color(255, 255, 255));
         viewsPanel.setPreferredSize(new java.awt.Dimension(885, 552));
         viewsPanel.setLayout(new java.awt.CardLayout());
-        viewsPanel.add(monitoringView1, "card2");
 
         javax.swing.GroupLayout mainPanelLayout = new javax.swing.GroupLayout(mainPanel);
         mainPanel.setLayout(mainPanelLayout);
@@ -90,7 +88,6 @@ public class WindowMain extends javax.swing.JFrame {
     private lagatrix.client.gui.components.complex.fragment.Header header;
     private lagatrix.client.gui.components.simple.RoundPanel mainPanel;
     private lagatrix.client.gui.components.complex.fragment.Menu menu;
-    private lagatrix.client.gui.views.main.MonitoringView monitoringView1;
     private javax.swing.JPanel viewsPanel;
     // End of variables declaration//GEN-END:variables
 }